modified folder structure; added comments
